% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{char}
\alias{char}
\title{Character vectors}
\usage{
char(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The object \code{x} coerced to an object of type "character".
}
\description{
This is a wrapper for \code{as.character()}.
}
\examples{
char(c(1, 2, 3))
}
