% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{classes}
\alias{classes}
\title{Variable classes}
\usage{
classes(df)
}
\arguments{
\item{df}{An input dataframe.}
}
\value{
A named character vector where names are the column names in \code{df} and values are the corresponding \code{class()} for each column.
}
\description{
Display the \code{class()} of each variable in a dataframe.
}
