% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{comNames}
\alias{comNames}
\title{Common column names}
\usage{
comNames(df1, df2)
}
\arguments{
\item{df1}{First input dataframe.}

\item{df2}{Second input dataframe.}
}
\value{
A character vector containing all of the column names that \code{df1} and \code{df2} have in common.
}
\description{
Get the common column names between two dataframes.
}
