% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{d}
\alias{d}
\title{Length of object dimensions}
\usage{
d(x)
}
\arguments{
\item{x}{An input object whose dimensions will be computed.}
}
\value{
Returns a length-\code{n} integer vector where \code{n} is the number of dimensions in \code{x} (e.g. \code{n=1} for atomic vectors and lists, \code{n=2} for matrices and dataframes, and \code{n} may be more than 2 for some arrays). Each element in the output is the number of elements in that dimension of \code{x} (e.g. if \code{n=1}, the number of elements in the list or vector, if \code{n=2}, first the number of rows in the dataframe or matrix, then the number of columns).
}
\description{
A common function to compute the number of elements (in vectors and lists) or the number of vectors in each dimension (dataframes, matrices, and arrays) of common types of \code{R} objects.
}
