% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{even}
\alias{even}
\title{Even numbers}
\usage{
even(x)
}
\arguments{
\item{x}{An object for which a method has been defined, or a numeric vector.}
}
\value{
A logical object with the same length as \code{x} indicating whether each element in x is odd or even
}
\description{
Determine whether values are odd or even.
}
