% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-packages.R
\name{find_packages}
\alias{find_packages}
\title{Find R package depedendencies}
\usage{
find_packages(path = "./")
}
\arguments{
\item{path}{Folder in which to search}
}
\description{
Recursively search R and Rmd files for packages that are loaded via `library()``
or `require()`.
}
\details{
This function will search for all files ending in ".R", ".r", or
".Rmd" and extract the subjects of library and require calls in them. The
intended use is to make it easier to identify package dependencies for a
project using R.
}
\examples{
\dontrun{
# Find packages and write to a txt file
packs <- find_packages("./R")
writeLines(packs, "required-packages.txt")
}
}
