% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-csv-excel-dates.R
\name{fix_csv_excel_dates}
\alias{fix_csv_excel_dates}
\title{Fix mangled dates}
\usage{
fix_csv_excel_dates(x)
}
\arguments{
\item{x}{([base::character()]) \cr
A character vector of dates}
}
\value{
A [base::Date()] vector.
}
\description{
Try to fix dates in a CSV file that has received the Excel kiss of death.
}
\examples{
x <- c("1/20/53", "1890-01-23", "12/1/53", "1/12/53", "1/1/09")
data.frame(input = x, output = fix_csv_excel_dates(x), stringsAsFactors = FALSE)

# decades/years less than or requal to today's are assumed to be this
# century, 20th century otherwise
now <- as.integer(substr(Sys.Date(), 3, 4))
x <- sprintf("1/1/\%s", now + -1:2)
data.frame(input = x, output = fix_csv_excel_dates(x), stringsAsFactors = FALSE)

}
