% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{medNA}
\alias{medNA}
\title{Median}
\usage{
medNA(x)
}
\arguments{
\item{x}{An object for which a method has been defined, or a numeric vector containing the values whose median is to be computed.}
}
\value{
Returns a length-one object of the same type as \code{x}, except when x is logical or integer of even length, when the result will be double.
}
\description{
This is a wrapper for \code{median(x, na.rm = TRUE)}.
}
