% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Plot objects to be plotted, separated by commas.}

\item{plotlist}{An alternate way to specify plot order, as a list. Defaults to \code{NULL}.}

\item{cols}{The number of columns in which the plots will be arranged. Defaults to 1.}

\item{layout}{An optional matrix of positions for each plot}
}
\value{
The plots in \code{...} or \code{plotlist} will be plotted in a clean manner. To further specify relative size, \code{windows()} can be called before \code{multiplot()}.
}
\description{
Plot multiple ggplot objects in the same well-formatted space.
}
