% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{namesLower}
\alias{namesLower}
\title{Change column names of a dataframe to lowercase}
\usage{
namesLower(df)
}
\arguments{
\item{df}{An input dataframe whose column names will be pushed to lowercase}
}
\value{
Returns the input dataframe, but with the names in lower case
}
\description{
Allows this operation to be done in a magrittr chain
}
