% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{nas}
\alias{nas}
\title{NA elements}
\usage{
nas(x)
}
\arguments{
\item{x}{An input vector.}
}
\value{
A length-one integer object displaying the number of NAs in \code{x}.
}
\description{
Counts the number of NA values in a vector.
}
