% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{num}
\alias{num}
\title{Numeric vectors}
\usage{
num(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The object \code{x} coerced to an object of type "numeric".
}
\description{
This is a wrapper for \code{as.numeric()}.
}
\examples{
num(c(1, 2, 3))
}
