% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predData.R
\name{predData}
\alias{predData}
\title{Model predictions}
\usage{
predData(mod, df, variable, vary = NULL, dv, type = "logit")
}
\arguments{
\item{mod}{A model whose coefficients and variance-covariance matrix will be used to do the predictions. Currently supports logistic regression (\code{glm} package) and hurdle models (\code{pscl} package).}

\item{df}{A dataframe containing data to be used in prediction. Should contain values for all variables used in \code{mod}.}

\item{variable}{A string specifying the variable that will be varied from its minimum to maximum values.}

\item{vary}{An (optional) string specifying a variable to interact with \code{variable} in doing the predictions. Defaults to \code{NULL}.}

\item{dv}{A string specifying which of the variables in \code{df} is the dependent variable in \code{mod}.}

\item{type}{A string specifying the model type (either 'logit' or 'hurdle'). Defaults to 'logit'.}
}
\value{
A dataframe containing values of \code{variable} (as well as the name of the variable) and corresponding predicted values, as well as the values of the boundaries of 95 percent intervals. These values can easily be used to plot the substantive effect of \code{variable} (and \code{vary}, if applicable) on \code{dv}, according to the results of \code{mod}.
}
\description{
Get data for model predictions and uncertainty varied over the range of \code{variable} while holding all other variables at their median values. Can support two-way interactions by specifying a variable name for \code{vary}.
}
