% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{size}
\alias{size}
\title{Object size}
\usage{
size(x, unit = "Mb")
}
\arguments{
\item{x}{An \code{R} object.}

\item{unit}{A string containing the units to be used. Defaults to "Mb". See \code{?object.size} for more options.}
}
\value{
An estimate of the size of the object in the units specified by \code{unit}. Estimates are not always accurate, see \code{?object.size} for details.
}
\description{
Provides an estimate of the memory that is being used to store an \code{R} object. This is a wrapper for \code{print(object.size(x), units=unit)}.
}
