% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{stdz}
\alias{stdz}
\title{Standardized variables}
\usage{
stdz(x)
}
\arguments{
\item{x}{An object of type numeric, logical, date, date-time, or time interval.}
}
\value{
The object \code{x} coerced to an object of type "numeric".
}
\description{
This function standardizes variables by taking the difference between each value and the mean and dividing by the standard deviation.
}
