% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{str_delete}
\alias{str_delete}
\title{Remove elements of a vector matching a pattern from a string}
\usage{
str_delete(string, pattern)
}
\arguments{
\item{string}{Input vector. either a character vector or something coercible to one.}

\item{pattern}{Pattern to look for.}
}
\value{
A vector of the same type as \code{string}, but without any elements in which \code{pattern} was detected.
}
\description{
Vectorised over \code{string} but not over \code{pattern}.
}
