% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{thead}
\alias{thead}
\title{Tibble Head}
\usage{
thead(x)
}
\arguments{
\item{x}{A dataframe}
}
\value{
The first six rows of object \code{x} coerced to an object of type "tbl_df".
}
\description{
This is a wrapper for \code{head()}.
}
\examples{
df <- data.frame(x=1:10, y=rnorm(10))
thead(df)
}
